separate components in checkout pages
