+++ client/src/components/Checkout-List/createData.js
@@ -1,0 +1,5 @@
+const createData = (productImg, productName, productQuantity, productTotalPrice) => {
+  // const density = population / size;
+  return { productImg, productName, productQuantity, productTotalPrice };
+};
+module.exports = createData;
